override Waffle middleware instead of convoluted view changes
